Ticket: #3743
Completed commit.  Added cancellation / shutdown support to the ServiceLock processor. Similar to the WorkQueue code changes, introduced a new ServiceLockThreadPool, which keeps track of the service lock entries, and also cancels them when appropriate.  Also introduced into each of the ServiceLock processors the ability to cancel in the middle.  Note that when a reinventory or reprocess is canceled, both of these operations reschedule themselves and start up again running the entire reinventory/reprocess.

+++ ImageServer/Services/ServiceLock/FilesystemDelete/FilesystemDeleteItemProcessor.cs
@@ -169,7 +169,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
     /// <summary>
     /// Class for processing 'FilesystemDelete' <see cref="Model.ServiceLock"/> rows.
     /// </summary>
-    public class FilesystemDeleteItemProcessor : BaseServiceLockItemProcessor, IServiceLockItemProcessor
+    public class FilesystemDeleteItemProcessor : BaseServiceLockItemProcessor, IServiceLockItemProcessor, ICancelable
     {
         #region Private Members
 		private DateTime _scheduledTime = Platform.Time;
@@ -264,7 +264,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 
 			foreach (FilesystemQueue queueItem in candidateList)
             {
-                if (_bytesToRemove < 0)
+				if (_bytesToRemove < 0 || CancelPending)
                     return;
 
 				StudyProcessStatistics stats = new StudyProcessStatistics("DeleteStudy");
@@ -344,7 +344,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 
 			foreach (FilesystemQueue queueItem in candidateList)
 			{
-				if (_bytesToRemove < 0)
+				if (_bytesToRemove < 0 || CancelPending)
 					break;
 
 				StudyProcessStatistics stats = new StudyProcessStatistics("PurgeStudy");
@@ -429,7 +429,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 			FilesystemProcessStatistics summaryStats = new FilesystemProcessStatistics("FilesystemTierMigrateInsert");
         	foreach (FilesystemQueue queueItem in candidateList)
         	{
-        		if (_bytesToRemove < 0)
+				if (_bytesToRemove < 0 || CancelPending)
         		{
                     Platform.Log(LogLevel.Debug, "Estimated disk space has been reached.");
                     break;
@@ -532,6 +532,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
                 if (list.Count > 0)
                 {
                     ProcessStudyDeleteCandidates(list);
+					if (CancelPending) break;
                 }
                 else
                 {
@@ -562,6 +563,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 				if (list.Count > 0)
 				{
 					ProcessStudyPurgeCandidates(list);
+					if (CancelPending) break;
 				}
 				else
 				{
@@ -594,6 +596,7 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 				if (list.Count > 0)
                 {
                     ProcessStudyMigrateCandidates(list);
+					if (CancelPending) break;
                 }
                 else
                 {
@@ -736,10 +739,10 @@ namespace ClearCanvas.ImageServer.Services.ServiceLock.FilesystemDelete
 
                         MigrateStudies(item, fs);
 
-                        if (_bytesToRemove > 0)
+                        if (_bytesToRemove > 0 && !CancelPending)
                             DeleteStudies(item, fs);
 
-                        if (_bytesToRemove > 0)
+                        if (_bytesToRemove > 0 && !CancelPending)
                             PurgeStudies(item, fs);
